+++ CHANGELOG.rst
@@ -31,6 +31,14 @@ Change Log
 - [???] "asynch" multienv
 - [???] properly model interconnecting powerlines
 
+
+[1.9.7] - 20xx-yy-zz
+----------------------
+- [IMPROVED] the function to check the backend interface now also check that
+  the `topo_vect` returns value between 1 and 2.
+- [IMPROVED] the function to check backend now also check the `topo_vect`
+  for each type of elements.
+
 [1.9.6] - 2023-10-26
 ----------------------
 - [BREAKING] when a storage is connected alone on a bus, even if it produces / absorbs 0.0 MW it